merge fix_mm to unstable

+++ memory_manager/mm_area.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   mm_area.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: trgaspar <trgaspar@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ekrebs <ekrebs@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/16 21:21:06 by ekrebs            #+#    #+#             */
-/*   Updated: 2025/02/17 16:26:44 by trgaspar         ###   ########.fr       */
+/*   Updated: 2025/02/17 17:37:45 by ekrebs           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,8 +41,25 @@ void	mm_area_create(t_mm *mm, char *new_area_name)
 void	mm_area_delete(t_mm *mm, char *area_name_to_delete)
 {
 	t_area_node	*area_node;
+	t_area_node *killme;
 
-	area_node = go_to_area_name(mm, area_name_to_delete);
-	destroy_area(area_node);
+	area_node = mm->areas;
+	if (ft_strcmp(mm->areas->area_name, area_name_to_delete) == EXIT_SUCCESS)
+	{
+		mm->areas = mm->areas->next;
+		destroy_area(area_node);
+		return ;
+	}
+	while(area_node->next)
+	{
+		if (ft_strcmp(area_node->next->area_name, area_name_to_delete) == EXIT_SUCCESS)
+		{
+			killme = area_node->next;
+			area_node->next = area_node->next->next;
+			destroy_area(killme);
+			return ;
+		}
+		area_node = area_node->next;
+	}
 	return ;
 }
